use if block for better meaning

+++ bin/ikill.sh
@@ -2,17 +2,19 @@
 ps_header=--headers
 for pid in `pgrep "$@"`
 do
-    ps -l $ps_header $pid || continue
-    echo "kill? (Y/n/term/kill/stop/...signal)"
-    read input
-    case "$input" in
-        n*|N*) signal=0 ;;
-        y*|Y*) signal=term ;;
-        "") signal=term ;;
-        *) signal="$input" ;;
-    esac
+    if ps -l $ps_header $pid
+    then
+        echo "kill? (Y/n/term/kill/stop/...signal)"
+        read input
+        case "$input" in
+            n*|N*) signal=0 ;;
+            y*|Y*) signal=term ;;
+            "") signal=term ;;
+            *) signal="$input" ;;
+        esac
 
-    kill -s $signal $pid
+        kill -s $signal $pid
+    fi
 
     ps_header=--no-headers
 done
